refactor(askiris): extract the segment state machine, tighten the scroll follow
useConversationSegments takes over archived, the segment id, the latest-messages
ref and startNewSegment, and ThreadItem moves with it. It sits next to the chat
rather than in src/hooks: that directory holds either shared feature logic or
single-caller mechanism, and this is neither. The chat keeps the two places that
decide when a boundary happens — the mount-switch effect and the composer button
— which is the line worth drawing: what a boundary does versus what triggers one.

The scroll follow stays inline. It was briefly a useStickToBottom hook, but the
hard half of the problem is knowing that the content changed, and a hook cannot
answer that for a caller — it can only take a signal and trust it. What was left
was a wrapper that looked reusable while handing the difficult part back. Inline,
it also keeps the listener on React's onScroll, where the element being absent
during the empty state is a non-issue. The follow now keys on the rendered thread
rather than the raw messages, so it tracks what is actually on screen.

AskIrisChat 246 -> 205 lines.

Two behaviour fixes. A segment boundary no longer clears the composer: losing
half-typed text is hostile, and a mount switch from the nav did it without the
user touching the chat at all. And the "new topic" button is gated on the live
thread rather than on what is rendered, so a replayed fixture no longer makes it
clickable with nothing to archive.

Verified against the dev server, mid-stream: scrolled to the top and content grew
1200 -> 2582px with the view staying put; scrolling back to the bottom resumed
following (3814 -> 4118px, ending flush); "new topic" from the top of a 4118px
thread jumped to the fresh divider; typed text survived the boundary.
